+++ Backend/src/main/java/com/codecool/backend/controller/dto/UserEntityDto.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.Set;
 
-public record UserEntityDto(int id, String name, String email, BigDecimal target, Set<Transaction> transactions) {
+public record UserEntityDto(Long id, String name, String email, BigDecimal target, Set<Transaction> transactions) {
     public UserEntityDto(UserEntity userEntity) {
         this(userEntity.getId(), userEntity.getName(), userEntity.getEmail(), userEntity.getTargetAmount(), userEntity.getTransactions());
     }
